Fix process killing when start time has small skews. [r=radix,jkakar,therve] [f=257070]
The way we obtain boot time may mean that a different timestamp is
returned on subsequent calls. As a consequence, processes may also get
a slightly imprecise start time. The process killing plugin would
verify that the start time matched precisely, so it might blow up
given the right circumstances.

Instead of generating a stable boot time, which is non-trivial business
when using /proc/uptime, we just accept small skews in the plugin, as
Chris suggested.  Woohay simplicity!