Merged add-vminfo-to-computer-info-message [r=therve,free].
The landscape client now reports whether it is running on a virtual machine or not.
For kvm checks, it wasn't possible to assume that /proc/cpuinfo will always exist because it would force all tests to create that file, even if it wasn't a specific test for it, as discussed with free.