Ignore to report packages that are coming from backports.
The reason for this is that backports are enabled by default from
xenial, but landscape doesn't handle that well, since we don't support
pinning properly.

If we report packages from backports, it means that Landscape will
upgrade to those version if you use an upgrade profile or select all
upgrades to their latest versions.