Allow bad chars in system vendor files
And while we're at it, clean up some of the mess in fs.py: there's
almost no circumstances where we want landscape client to fall over
because a text file has invalid UTF-8 (the original function before the
py2/3 split had no such circumstances that'd lead to decoding errors).
Furthermore, read_binary_file's limit was wrong when positive (not that
anything called it with a positive limit) and read_text_file defeated
the purpose of having a limit by permitting an arbitrarily large read
even when one was set.