Add Ubuntu Pro Info monitor (#118)
This patch adds a monitor for the Ubuntu pro status of the client machine. This status is sent to the Landscape Server as a single string of JSON, where it is parsed, stored, and presented to the user. Pro status is checked at client registration, and is updated every 15 minutes.