Read temperature from hwmon devices
On most systems hwmon provides access to the temperature information on
more devices, including the CPU and GPU on most systems. These
additional temperature sources yield a more accurate maximum temperature
for the overall system.

With this change, the following precedence is used for reading thermal
zones (depending on availability):
1. "/sys/class/hwmon/*/temp*_input"
2. "/sys/class/thermal/*/temp"
3. "/proc/acpi/thermal_zone/*/temperature"