fix: get tests passing on plucky (#290)
Twisted changed a little bit - it used to run `cleanFailure` on
`Failure` instances in the `defer` code but no longer does in the
version available in plucky (python3-twisted 24.10).

This change does `cleanFailure` in the `FakeReactor` we use for some
tests. This allows tests that rely on catching and flushing `Failure`s
to succeed.